services/ai_processor/requirements.txt
psycopg2-binary
redis
sqlalchemy
openai
python-dotenv
fpdf2
jinja2
